DTSPB-909 - Add new pages to e2es
