+++ src/test/end-to-end/paths/solicitor/applyGrantOfProbateMultiExecutor.js
@@ -65,6 +65,7 @@ Scenario('Solicitor - Apply Grant of probate Multi Executor', async function (I)
     await I.grantOfProbatePage2();
     await I.grantOfProbatePage3();
     await I.grantOfProbatePage4();
+    await I.grantOfProbatePage5();
     await I.cyaPage();
 
     await I.seeEndState(endState);
